+++ connection.go
@@ -211,7 +211,8 @@ type Connection struct {
 	// requestCnt is a counter of active requests.
 	requestCnt int64
 
-	pool allocator
+	// alloc is an allocator for response buffers.
+	alloc Allocator
 }
 
 var _ = Connector(&Connection{}) // Check compatibility with connector interface.
@@ -329,9 +330,9 @@ type Opts struct {
 	Handle interface{}
 	// Logger is user specified logger used for error messages.
 	Logger Logger
-	// PoolSizes is a list of pool sizes (power-of-two) for each connection,
-	// nil value means default -- not using pool inside.
-	PoolSizes []int
+	// Allocator is used to allocate and deallocate response buffers.
+	// If nil, default Go allocation is used.
+	Allocator Allocator
 }
 
 // Connect creates and configures a new Connection.
@@ -378,7 +379,12 @@ func Connect(ctx context.Context, dialer Dialer, opts Opts) (conn *Connection, e
 		conn.opts.Logger = defaultLogger{}
 	}
 
+	if conn.opts.Allocator != nil {
+		conn.alloc = conn.opts.Allocator
+	}
+
 	conn.cond = sync.NewCond(&conn.mutex)
+
 	if conn.opts.Reconnect > 0 {
 		// We don't need these mutex.Lock()/mutex.Unlock() here, but
 		// runReconnects() expects mutex.Lock() to be set, so it's
@@ -398,10 +404,6 @@ func Connect(ctx context.Context, dialer Dialer, opts Opts) (conn *Connection, e
 	go conn.pinger()
 	if conn.opts.Timeout > 0 {
 		go conn.timeouts()
-	}
-
-	if conn.opts.PoolSizes != nil {
-		conn.pool = newPooler(conn.opts.PoolSizes)
 	}
 
 	// TODO: reload schema after reconnect.
@@ -891,7 +893,7 @@ func (conn *Connection) reader(r io.Reader, c Conn) {
 	for atomic.LoadUint32(&conn.state) != connClosed {
 		var err error
 
-		buf, err = read(r, conn.lenbuf[:], conn.pool)
+		buf, err = read(r, conn.lenbuf[:], conn.alloc)
 
 		if err != nil {
 			err = ClientError{
@@ -1225,18 +1227,20 @@ func (conn *Connection) timeouts() {
 	}
 }
 
-// read uses args to allocate slices for responses using sync.Pool.
-// data must be released later using Release.
-func read(r io.Reader, lenbuf []byte, pool allocator) (buf smallBuf, err error) {
+// read reads a response from the reader and allocates a buffer using the
+// allocator. If alloc is nil, a regular Go allocation is used.
+// The returned buffer could be released using Release.
+func read(r io.Reader, lenbuf []byte, alloc Allocator) (smallBuf, error) {
 	var length uint64
 
-	if _, err = io.ReadFull(r, lenbuf); err != nil {
-		return
+	if _, err := io.ReadFull(r, lenbuf); err != nil {
+		return smallBuf{}, fmt.Errorf("failed to read response length: %w", err)
 	}
+
 	if lenbuf[0] != 0xce {
-		err = errors.New("wrong response header")
-		return
+		return smallBuf{}, errors.New("wrong response header")
 	}
+
 	length = (uint64(lenbuf[1]) << 24) +
 		(uint64(lenbuf[2]) << 16) +
 		(uint64(lenbuf[3]) << 8) +
@@ -1244,17 +1248,19 @@ func read(r io.Reader, lenbuf []byte, pool allocator) (buf smallBuf, err error) 
 
 	switch {
 	case length == 0:
-		err = errors.New("response should not be 0 length")
-		return
+		return smallBuf{}, errors.New("response should not be 0 length")
 	case length > math.MaxUint32:
-		err = errors.New("response is too big")
-		return
+		return smallBuf{}, errors.New("response is too big")
 	}
 
-	buf = CreateBuf(pool, int(length))
-	_, err = io.ReadFull(r, buf.b)
+	buf := createBuf(alloc, int(length))
+	if _, err := io.ReadFull(r, buf.b); err != nil {
+		buf.Release()
 
-	return
+		return smallBuf{}, fmt.Errorf("failed to read response: %w", err)
+	}
+
+	return buf, nil
 }
 
 func (conn *Connection) nextRequestId(context bool) (requestId uint32) {