+++ smallbuf.go
@@ -6,10 +6,10 @@ import (
 )
 
 type smallBuf struct {
-	b    []byte
-	p    int
-	ptr  *[]byte
-	pool allocator
+	b     []byte
+	p     int
+	ptr   *[]byte
+	alloc Allocator
 }
 
 func (s *smallBuf) Read(d []byte) (l int, err error) {
@@ -68,18 +68,22 @@ func (s *smallBuf) Seek(offset int) error {
 	return nil
 }
 
-func CreateBuf(a allocator, size int) smallBuf {
+func createBuf(a Allocator, size int) smallBuf {
 	if a == nil {
 		return smallBuf{b: make([]byte, size)}
 	}
 
-	ptr := a.getSlice(size)
-	return smallBuf{ptr: ptr, pool: a, b: *ptr}
+	ptr := a.Get(size)
+	if ptr == nil {
+		return smallBuf{b: make([]byte, size)}
+	}
+
+	return smallBuf{b: *ptr, ptr: ptr, alloc: a}
 }
 
 func (s *smallBuf) Release() {
-	if s.pool != nil && s.ptr != nil {
-		s.pool.putSlice(s.ptr)
+	if s.alloc != nil && s.ptr != nil {
+		s.alloc.Put(s.ptr)
 	}
 }